Enhance TranslatedResults functionality
With the advent of the query results retriever some views can actually be added
only to the default locale page but still return results appropriate to the
locale of whatever page the user is viewing.  Because of this we did not want
to have View->TranslatedResults calling getTranslation on every result when
some of these results are already in the correct translation.

To accomplish this we made two changes:

 1. Move TranslatedResults function to ViewResultsRetriever.  This allows it to
    be overridden by other ResultsRetriever types as necessary.
 2. Change the default implementation to first check if the result is already
    in the correct locale before trying to get the translation.

We also removed the disable_locale_filter from HandPickedResultsRetriever and
instead put it in ViewResultsRetriever->TranslatedResults to disable it for
all calls to ViewResultsRetriever->Results from the TranslatedResults function