Refactor how special values are added to FPV
FieldPredicateValue has a mechanism for "special" values that can not be
defined explicitly at the time of view/query creation.  For instance, you may
want to query based on children of "this page".  But a view can be defined on
one page and used on all its children (and translations, and children's
translations, etc).  Thus, you need to be able to save a query with a special
value for your query that will dynamically look up the page ID of the current
page when the query is executed so that it works in the context of the page
that the query is actually run from.

This commit refactors those special values from being something that is hard-
coded to something that can be plugged in via configuration.