Add predicate conditions and change how tokens work
This commit introduces two big changes:

1. There is now a class of objects called "predicate conditions" that can be
applied to all QueryPredicate instances.  These conditions allow you to build
queries that do not always contain every predicate defined in the query.  In
our canonical example, we imagine that we have a FieldPredicate instance that
uses a query string parameter as its value (for instance, "Category = $param").
In order to have a single page that shows things from all categories, but then
only shows things from a single category when the category URL parameter is
present, you need the "Category = $param" FieldPredicate to only be added to
the QueryBuilder object *if* you actually have the URL param defined.  By
adding a QueryParamPredicateCondition to your QueryPredicate you can now do
this.

2. The way that "special values" were handled was refactored to be more
flexible.  Previously an entire value must be *only* the special value key.
Additionally, there were no parameters to the special values.  This meant that
if you wanted to use a query parameter as a special value you had to define
a new special value for every query parameter you'd ever use.  Now special
values can be placed in the middle of a FieldPredicateValue string so that they
can be used in "like" predicates, etc.  They can also have parameters so that a
single special value (now called value tokens) can be registered to handle all
query parameters.  So, you could have a FieldPredicateValue with a value that
is "$$QueryParam:year$$-$$QueryParam:month$$%' and use it in a "like" clause on
a date column.  It would use both query parameters to build the actual value
used by the SQL statement.