Add support for querying based on a translation of the current page
Added a new token value that can be used to get a translation of the current
page based on a URL param that contains the desired locale name.

Also made QueryResultsRetriever disable the locale filter when it is querying
for results since it can query for results in another language.

Also provided helper functions for testing if a view has results without using
TranslatedResults since in same cases you have a view that queries things in
another locale and you don't want the translation.