Fix the bad relationships between ViewHost and View
SilverStripe doesn't support a has_many relationship on a DataObjectDecorator.
We previously had faked it, but by doing so broke other things.  We also made
it so that only SiteTree could ever have ViewHost added to it since we were
looking up the has_many ourselves based on "HostID" and assuming SiteTree.

Now the ViewHost DOD has a has_one to ViewCollection, and ViewCollection in
turn has the has_many to View.  This works with the SS built-in functionality
and made it so that we could remove a couple of ugly hacks we had before.