Fix how compound where clauses are built
CompoundPredicate was using an internal query to build compound where clauses,
but this lost all joins, sorts, etc that might be added by its child predicates
since it was only copying the 'wheres' SQL part over.  Copying joins would be
problematic.

So, now QueryBuilder supports the starting and stopping of compound where
clause groups and this is used by CompoundPredicate.