Fix exclusive use of TaxonomyTermPredicate
When used with Inclusive == false, this predicate was returning false positives
(that is, nodes that actually DID have the taxonomy term they were supposed to
exclude) because the LEFT OUTER JOIN to SiteTree_VocabularyTerms and
VocabularyTerms was causing them to get included.  The LEFT OUTER JOIN is
needed, but can only be to a single table.  Besides, the extra join to
VocabularyTerm is really superfluous anyway since we already know the term
exists since the predicate itself is joined to the term and has retrieved it by
ID.