Fix the original problem 6daac10b tried to fix
The problem was that a TaxonomyTermPredicate could not be used in a
CompoundPredicate that was exclusive (OR instead of AND) because it was using
an inner join for the inclusive criteria.

This commit fixes that issue by using a left outer join and ensuring that the
joined table actually contains a joining row by using a WHERE "STVT".ID IS NOT
NULL.

This fixes the original intent of 6daac10b and does not break the exclusive
case by just using "IS NULL" instead of "<>".