Fix bug in ownership of view causing bad links
If you have a page that is Language X, and that page does not contain
a view, but the default locale (Language D) version of the page does,
and this view is returned from GetView, and then ->Link() is used
to get the RSS link while on page X, the link could be created to the
page D URL if D had a ViewCollectionID, but X had zero as its VCID.

Essentially, if a view is found on a particular page, it should act as
if it is owned by that page for the purpose of creating links to the view
(the only place that the view's owner is currently being used) so that
links are created relative to the page that found the view, not the page
that actually contains the view in its view collection.  This ensures that
the link is to the correct translation, and that later on a view could be
added to that translation and the link would still work, allowing you to
start with a single view on your default locale, but later override that
view on a per-locale basis.