Add more qualifiers and raw SQL functionality
We needed more qualifiers like >, >=, <, <=.

Also, at times a value may be a raw SQL string - for instance, a SQL
function like DATE_SUB(NOW(), INTERVAL 30 DAY).  In those instances
you do not want to add tick marks around the value.  We added a boolean
that allows you to specify if the value(s) for a FieldPredicate are raw
SQL.