Change defaults per results retr. type
The default for all results retrievers is none since each type works
differently.  However, the default for hand-picked results retrievers
is to translate into the locale of the page since the most common use
pattern is to set the view on the page of your master language, but then
have it translate the results into the page it is being used on.