Changes order of JOINs in HandPickedResultsRetriver for performance.
The MySQL query optimizer was performing JOINs in an ineffcient order. Changing
the order of the INNER JOIN to SiteTree_translationgroups so that it happens
before the LEFT JOINs to the master SiteTree node provides enough of a hint to
the optimizer to do that join first. This reduces query execution time significantly.