Prevents QueryBuilderField from entering infinite loop
QueryBuilderField would enter an infinite recursive loop on some classes when
traverse_has_one was set to true. This was introduced with the change to using
the Config class, since by default it config inheritance (whereas the original
implementation) relied on access statics on an individual class, but not it's
parents. Setting the Config::UNINHERITED flag fixes this.