Fixes Views Delete Functionality
In the Views tab (CMS Admin) of and SiteTree object, the delete button resulted
in a blank page being displaed to the user and the view not actually being
deleted, even though the server's response was 200 OK. This was a result of
using a bad field name in the GridField relation editor in ViewHost.php.
Changing this form 'ViewCollection.Views' to simply 'Views' fixed the problem.
View deletion now works as expected.