Stops Views Module from breaking i18nTextCollectorTask
Dyanmically building keys for the _t() function breaks the static analysis done
by i18nTextCollectorTask. This happens because the task doesn't actually
evaluate the expression. Instead it tries to use the whole, raw, unevaluated
expression as the literal key. This in turn breaks the array unflattener in
i18nTextCollector, causing it to try to assign a string index by reference,
which promptly makes PHP segfault. Getting rid of the constants fixes all of
this.