Adds user-adjustable sorting to aggregated views
Adding this feature required reworking how some aspects of how the
QueryBuilderField worked. Instead of using QueryBuilderField to select views for
the ViewAggregatingResultsRetriever, we now use a GridField (similar to how
HandPickedResultsRetriever works) to select and sort views. Internally this user
an extra SortOrder column on the intersection table.

This change alloed QueryBuilderField to be simplified quite a bit since it no
longer had to support many_many relationships.