Fixes pagination of results from a ViewAggregatingResultsRetriever
ViewAggregatingResultsRetriever was calling View::Results() on each of it's
views to build a coalesced result list. This was flawed, though, because
View::Results contains pagination logic wrapping around the call for
ViewResultsRetriever::results(). By making ViewAggregatingResultsRetriever
call ViewResultsRetriever::results() directly, we bypass this, making pagination
of the final result amalgam work as expected.